+++ pluggable-architecture/git-data-fetcher/git_stats/providers/base.py
