+++ pluggable-architecture/git-data-fetcher/git_stats/providers/github_provider.py
