+++ pluggable-architecture/git-data-fetcher/git_stats/providers/gitlab_provider.py
