+++ pluggable-architecture/git-data-fetcher/tests/test_download.py
@@ -1,4 +1,4 @@
-from git_data_fetcher.download import GitFetcher
+from git_stats.download import GitFetcher
 
 
 def test_url_parser():
